dizaynni o'zgartirishdan oldingisi

+++ config/jwt.php
@@ -10,9 +10,9 @@ return [
         'passphrase' => env('JWT_PASSPHRASE'),
     ],
 
-    'ttl' => env('JWT_TTL', 60),
+    'ttl' => (int) env('JWT_TTL', 60),
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
     'algo' => env('JWT_ALGO', 'HS256'),
 
@@ -29,11 +29,11 @@ return [
 
     'lock_subject' => true,
 
-    'leeway' => env('JWT_LEEWAY', 0),
+    'leeway' => (int) env('JWT_LEEWAY', 0),
 
     'blacklist_enabled' => env('JWT_BLACKLIST_ENABLED', true),
 
-    'blacklist_grace_period' => env('JWT_BLACKLIST_GRACE_PERIOD', 0),
+    'blacklist_grace_period' => (int) env('JWT_BLACKLIST_GRACE_PERIOD', 0),
 
     'decrypt_cookies' => false,
